use static_cast to address implicit conversion  warnings JIRA SB-3592
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@100101 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/algo/blast/proteinkmer/mhfile.hpp
@@ -161,7 +161,7 @@ public:
     void GetMinHits(int oid, int& subjectOid, vector<uint32_t>& hits) const;
 
     /// Returns the number of hash arrays.
-    int GetNumSignatures() const {return (m_DataFileSize/(GetDataWidth()*GetNumHashes()+4));}
+    int GetNumSignatures() const {return static_cast<int>(m_DataFileSize/(GetDataWidth()*GetNumHashes()+4));}
 
 private:
